src/ticket/ticket.service.spec.ts
import { Test, TestingModule } from '@nestjs/testing';
import { TicketService } from './ticket.service';
import { TicketRepository } from './ticket.repository';

describe('TicketService', () => {
  let service: TicketService;
  let mockTicketRepository;

  beforeEach(async () => {
    mockTicketRepository = {
      count: jest.fn(),
      get: jest.fn(),
    };

    const module: TestingModule = await Test.createTestingModule({
      providers: [
        TicketService,
        {
          provide: TicketRepository,
          useValue: mockTicketRepository,
        },
      ],
    }).compile();

    service = module.get<TicketService>(TicketService);
  });

  it('should be defined', () => {
    expect(service).toBeDefined();
  });

  describe('count', () => {
    it('should handle errors gracefully', async () => {
      mockTicketRepository.count.mockRejectedValue(new Error('Test error'));
      await expect(service.count({ /* invalid condition */ })).rejects.toThrow('Test error');
    });

    it('should return total count when called with empty conditions', async () => {
      mockTicketRepository.count.mockResolvedValue(100); // Assuming 100 is the total count
      const result = await service.count({});
      expect(result).toBe(100);
    });


    it('should return correct count for specific conditions', async () => {
      mockTicketRepository.count.mockResolvedValue(20); // Assuming 20 matches the condition
      const condition = { status: 'open' };
      const result = await service.count(condition);
      expect(result).toBe(20);
    });

    it('should return zero when no records match the condition', async () => {
      mockTicketRepository.count.mockResolvedValue(0);
      const condition = { status: 'closed' };
      const result = await service.count(condition);
      expect(result).toBe(0);
    });

    it('should handle null or undefined conditions', async () => {
      mockTicketRepository.count.mockResolvedValue(0); // Assuming no tickets for null/undefined condition
      await expect(service.count(null)).resolves.toBe(0);
      await expect(service.count(undefined)).resolves.toBe(0);
    });
  });

  describe('bar', () => {
    it('should handle errors gracefully', async () => {
      mockTicketRepository.get.mockRejectedValue(new Error('Test error'));
      await expect(service.bar({ /* invalid condition */ })).rejects.toThrow('Test error');
    });

    it('should return empty labels and series for empty result set', async () => {
      mockTicketRepository.get.mockResolvedValue([]);
      const [labels, series] = await service.bar({ /* conditions */ });
      expect(labels).toEqual([]);
      expect(series).toEqual([]);
    });

    it('should propagate errors from the repository', async () => {
      const errorMessage = 'Database error';
      mockTicketRepository.get.mockRejectedValue(new Error(errorMessage));
      await expect(service.bar({})).rejects.toThrow(errorMessage);
    });
  });
});
